Game.players map now uses uuid keys instead of ws
Player info needs to be sent back to client as pa-
rtof the game data. A key of type websocket object
cannot be stringified in a meaningful way, so ins-
tead of using the client's websocket, the uuid is
used.

This means the player data can be packaged as part
of the game data as an object with the client uuids
as the object keys, and the class fields as the
object properties. So the clients can use its uuid
to pick out data that pertains to that particular
client, or use all the data for all the players -
which will later be necessary for drawing the snakes
on screen and showing each player's score, stuff
like that.