When client disconnects, attempt to reconnect
This has exposed a fundamental flaw in the approach
to client metadata on the server. It was originally
based on the websocket object itself - but if a cl-
ient disconnects then reconnects, it will be a tot-
ally different object.

Therefore, the server will instead generate a uuid
and pass that to the client. If the client discon-
nects during the game, it will immediately try to
reconnect, providing the uuid so the server knows
which of the clients/players this relates to.