Package lobbyList to make easier to find particular lobbies
Used to be an array, now an object with lobby_names
for the keys. This can easily be turned into an array
on the client end with Object.values(), but with the
lobby_names as keys it makes it very easy to get a
single lobby that we are interested in instead of
having to iterate through an array to find a partic-
ular one.